asset-management-automation-py/requirements.txt
selenium>=4.0.0
pytest>=7.0.0
allure-pytest>=2.13.0
webdriver-manager>=4.0.0
pytest-html>=4.0.0
loguru>=0.7.0
pytest-xdist>=3.0.0
pytest-rerunfailures>=12.0
python-dotenv==1.0.1
requests==2.31.0
pillow==10.2.0 